# medi-care-application/generated/src/main/java/com/eaglecare/api/PatientsApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (6.0.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.medicare.api;

import java.math.BigDecimal;
import com.medicare.model.Patients;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2024-09-12T00:43:21.532621800+05:30[Asia/Calcutta]")
@Validated
@Tag(name = "Patients", description = "Operations about patients")
public interface PatientsApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * POST /v1/api/patients : Creation of new patients
     * Return the list values
     *
     * @param patients Update an existent pet in the store (required)
     * @return successful operation (status code 200)
     *         or Invalid Patients id (status code 400)
     */
    @Operation(
        operationId = "createNewPatients",
        summary = "Creation of new patients",
        tags = { "Patients" },
        responses = {
            @ApiResponse(responseCode = "200", description = "successful operation", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Patients.class))
            }),
            @ApiResponse(responseCode = "400", description = "Invalid Patients id")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/v1/api/patients",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    default ResponseEntity<Patients> _createNewPatients(
        @Parameter(name = "Patients", description = "Update an existent pet in the store", required = true) @Valid @RequestBody Patients patients
    ) {
        return createNewPatients(patients);
    }

    // Override this method
    default  ResponseEntity<Patients> createNewPatients(Patients patients) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"basic_info\" : { \"role\" : { \"name\" : \"ROLE_USER\", \"description\" : \"User role\", \"id\" : \"4acd0b30-0ac1-4d0f-938b-c7b86d804375\" }, \"gender\" : \"Male\", \"user_id\" : \"fb17eecf-5ec6-4c13-a5bd-5ce1451472d0\", \"last_name\" : \"Aravind\", \"employment_details\" : { \"joining_date\" : \"2024-11-05T00:00:00Z\", \"relieving_date\" : \"2024-11-05T00:00:00Z\", \"id\" : \"4acd0b30-0ac1-4d0f-938b-c7b86d804375\", \"payroll\" : { \"id\" : \"4acd0b30-0ac1-4d0f-938b-c7b86d804375\", \"gross_amount\" : 0.8008282, \"net_amount\" : 1.4658129, \"deductions\" : 6.0274563, \"payment_date\" : \"2024-11-05T00:00:00Z\" }, \"department\" : \"CLINIC_STAFF\", \"salary\" : 18500 }, \"first_name\" : \"Sastika\" }, \"contact\" : { \"address\" : { \"country\" : \"India\", \"geo_location\" : \"India\", \"pin\" : \"624601\", \"city\" : \"Coimabtore\", \"street\" : \"121,Balaji Nagar\", \"address_category\" : \"Personal\", \"id\" : \"4acd0b30-0ac1-4d0f-938b-c7b86d804375\", \"state\" : \"Tamilnadu\" }, \"phone\" : \"7010192967\", \"id\" : \"4acd0b30-0ac1-4d0f-938b-c7b86d804375\", \"email\" : \"aravind@gmail.com\" }, \"id\" : \"4acd0b30-0ac1-4d0f-938b-c7b86d804375\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * DELETE /v1/api/patients/{id} : Delete Operation
     * Delete patients account
     *
     * @param id User Id (required)
     * @return successful operation (status code 200)
     *         or Invalid user id (status code 400)
     */
    @Operation(
        operationId = "deletePatientsById",
        summary = "Delete Operation",
        tags = { "Patients" },
        responses = {
            @ApiResponse(responseCode = "200", description = "successful operation", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = String.class))
            }),
            @ApiResponse(responseCode = "400", description = "Invalid user id")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/v1/api/patients/{id}",
        produces = { "application/json" }
    )
    default ResponseEntity<String> _deletePatientsById(
        @Parameter(name = "id", description = "User Id", required = true) @PathVariable("id") String id
    ) {
        return deletePatientsById(id);
    }

    // Override this method
    default  ResponseEntity<String> deletePatientsById(String id) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "\"Doctor removed successfully\"";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /v1/api/patients : Get Patients
     * Get Patients list by params
     *
     * @param page Page number (required)
     * @param count Data count (required)
     * @param patientsId Patients Id (optional, default to 4598)
     * @param phoneNumber Patients Phone Number (optional, default to 7010182967)
     * @return successful operation (status code 200)
     *         or Invalid user id (status code 400)
     */
    @Operation(
        operationId = "getPatients",
        summary = "Get Patients",
        tags = { "Patients" },
        responses = {
            @ApiResponse(responseCode = "200", description = "successful operation", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Patients.class))
            }),
            @ApiResponse(responseCode = "400", description = "Invalid user id")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/v1/api/patients",
        produces = { "application/json" }
    )
    default ResponseEntity<List<Patients>> _getPatients(
        @NotNull @Parameter(name = "page", description = "Page number", required = true) @Valid @RequestParam(value = "page", required = true, defaultValue = "0") BigDecimal page,
        @NotNull @Parameter(name = "count", description = "Data count", required = true) @Valid @RequestParam(value = "count", required = true, defaultValue = "5") BigDecimal count,
        @Parameter(name = "patientsId", description = "Patients Id") @Valid @RequestParam(value = "patientsId", required = false, defaultValue = "4598") BigDecimal patientsId,
        @Parameter(name = "phoneNumber", description = "Patients Phone Number") @Valid @RequestParam(value = "phoneNumber", required = false, defaultValue = "7010182967") String phoneNumber
    ) {
        return getPatients(page, count, patientsId, phoneNumber);
    }

    // Override this method
    default  ResponseEntity<List<Patients>> getPatients(BigDecimal page, BigDecimal count, BigDecimal patientsId, String phoneNumber) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"basic_info\" : { \"role\" : { \"name\" : \"ROLE_USER\", \"description\" : \"User role\", \"id\" : \"4acd0b30-0ac1-4d0f-938b-c7b86d804375\" }, \"gender\" : \"Male\", \"user_id\" : \"fb17eecf-5ec6-4c13-a5bd-5ce1451472d0\", \"last_name\" : \"Aravind\", \"employment_details\" : { \"joining_date\" : \"2024-11-05T00:00:00Z\", \"relieving_date\" : \"2024-11-05T00:00:00Z\", \"id\" : \"4acd0b30-0ac1-4d0f-938b-c7b86d804375\", \"payroll\" : { \"id\" : \"4acd0b30-0ac1-4d0f-938b-c7b86d804375\", \"gross_amount\" : 0.8008282, \"net_amount\" : 1.4658129, \"deductions\" : 6.0274563, \"payment_date\" : \"2024-11-05T00:00:00Z\" }, \"department\" : \"CLINIC_STAFF\", \"salary\" : 18500 }, \"first_name\" : \"Sastika\" }, \"contact\" : { \"address\" : { \"country\" : \"India\", \"geo_location\" : \"India\", \"pin\" : \"624601\", \"city\" : \"Coimabtore\", \"street\" : \"121,Balaji Nagar\", \"address_category\" : \"Personal\", \"id\" : \"4acd0b30-0ac1-4d0f-938b-c7b86d804375\", \"state\" : \"Tamilnadu\" }, \"phone\" : \"7010192967\", \"id\" : \"4acd0b30-0ac1-4d0f-938b-c7b86d804375\", \"email\" : \"aravind@gmail.com\" }, \"id\" : \"4acd0b30-0ac1-4d0f-938b-c7b86d804375\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /v1/api/patients/{id} : Creation of new Patients
     * Return the list values
     *
     * @param id User Id (required)
     * @return successful operation (status code 200)
     *         or Invalid user id (status code 400)
     */
    @Operation(
        operationId = "getPatientsById",
        summary = "Creation of new Patients",
        tags = { "Patients" },
        responses = {
            @ApiResponse(responseCode = "200", description = "successful operation", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Patients.class))
            }),
            @ApiResponse(responseCode = "400", description = "Invalid user id")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/v1/api/patients/{id}",
        produces = { "application/json" }
    )
    default ResponseEntity<Patients> _getPatientsById(
        @Parameter(name = "id", description = "User Id", required = true) @PathVariable("id") String id
    ) {
        return getPatientsById(id);
    }

    // Override this method
    default  ResponseEntity<Patients> getPatientsById(String id) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"basic_info\" : { \"role\" : { \"name\" : \"ROLE_USER\", \"description\" : \"User role\", \"id\" : \"4acd0b30-0ac1-4d0f-938b-c7b86d804375\" }, \"gender\" : \"Male\", \"user_id\" : \"fb17eecf-5ec6-4c13-a5bd-5ce1451472d0\", \"last_name\" : \"Aravind\", \"employment_details\" : { \"joining_date\" : \"2024-11-05T00:00:00Z\", \"relieving_date\" : \"2024-11-05T00:00:00Z\", \"id\" : \"4acd0b30-0ac1-4d0f-938b-c7b86d804375\", \"payroll\" : { \"id\" : \"4acd0b30-0ac1-4d0f-938b-c7b86d804375\", \"gross_amount\" : 0.8008282, \"net_amount\" : 1.4658129, \"deductions\" : 6.0274563, \"payment_date\" : \"2024-11-05T00:00:00Z\" }, \"department\" : \"CLINIC_STAFF\", \"salary\" : 18500 }, \"first_name\" : \"Sastika\" }, \"contact\" : { \"address\" : { \"country\" : \"India\", \"geo_location\" : \"India\", \"pin\" : \"624601\", \"city\" : \"Coimabtore\", \"street\" : \"121,Balaji Nagar\", \"address_category\" : \"Personal\", \"id\" : \"4acd0b30-0ac1-4d0f-938b-c7b86d804375\", \"state\" : \"Tamilnadu\" }, \"phone\" : \"7010192967\", \"id\" : \"4acd0b30-0ac1-4d0f-938b-c7b86d804375\", \"email\" : \"aravind@gmail.com\" }, \"id\" : \"4acd0b30-0ac1-4d0f-938b-c7b86d804375\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * PUT /v1/api/patients/{id} : Creation of new patients
     * Return the list values
     *
     * @param id Patients Id (required)
     * @param patients Update an existent Patients (required)
     * @return successful operation (status code 200)
     *         or Invalid Patients id (status code 400)
     */
    @Operation(
        operationId = "updatePatients",
        summary = "Creation of new patients",
        tags = { "Patients" },
        responses = {
            @ApiResponse(responseCode = "200", description = "successful operation", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Patients.class))
            }),
            @ApiResponse(responseCode = "400", description = "Invalid Patients id")
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/v1/api/patients/{id}",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    default ResponseEntity<Patients> _updatePatients(
        @Parameter(name = "id", description = "Patients Id", required = true) @PathVariable("id") String id,
        @Parameter(name = "Patients", description = "Update an existent Patients", required = true) @Valid @RequestBody Patients patients
    ) {
        return updatePatients(id, patients);
    }

    // Override this method
    default  ResponseEntity<Patients> updatePatients(String id, Patients patients) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"basic_info\" : { \"role\" : { \"name\" : \"ROLE_USER\", \"description\" : \"User role\", \"id\" : \"4acd0b30-0ac1-4d0f-938b-c7b86d804375\" }, \"gender\" : \"Male\", \"user_id\" : \"fb17eecf-5ec6-4c13-a5bd-5ce1451472d0\", \"last_name\" : \"Aravind\", \"employment_details\" : { \"joining_date\" : \"2024-11-05T00:00:00Z\", \"relieving_date\" : \"2024-11-05T00:00:00Z\", \"id\" : \"4acd0b30-0ac1-4d0f-938b-c7b86d804375\", \"payroll\" : { \"id\" : \"4acd0b30-0ac1-4d0f-938b-c7b86d804375\", \"gross_amount\" : 0.8008282, \"net_amount\" : 1.4658129, \"deductions\" : 6.0274563, \"payment_date\" : \"2024-11-05T00:00:00Z\" }, \"department\" : \"CLINIC_STAFF\", \"salary\" : 18500 }, \"first_name\" : \"Sastika\" }, \"contact\" : { \"address\" : { \"country\" : \"India\", \"geo_location\" : \"India\", \"pin\" : \"624601\", \"city\" : \"Coimabtore\", \"street\" : \"121,Balaji Nagar\", \"address_category\" : \"Personal\", \"id\" : \"4acd0b30-0ac1-4d0f-938b-c7b86d804375\", \"state\" : \"Tamilnadu\" }, \"phone\" : \"7010192967\", \"id\" : \"4acd0b30-0ac1-4d0f-938b-c7b86d804375\", \"email\" : \"aravind@gmail.com\" }, \"id\" : \"4acd0b30-0ac1-4d0f-938b-c7b86d804375\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}