# medi-care-application/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Hospital Management System
  description: |-
    This helps to manage the hospital
  termsOfService: http://swagger.io/terms/
  contact:
    email: aravind@monkeeys.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.11
...
servers:
  - url: localhost:8080/hospital
...
paths:
...
  /v1/api/patients:
    post:
      tags:
        - Patients
      summary: Creation of new patients
      description: Return the list values
      operationId: createNewPatients
      requestBody:
        description: Update an existent pet in the store
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Patients'
        required: true
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Patients'
        '400':
          description: Invalid Patients id
    get:
      tags:
        - Patients
      summary: Get Patients
      description: Get Patients list by params
      operationId: getPatients
      parameters:
        - name: page
          in: query
          description: Page number
          required: true
          explode: true
          schema:
            type: number
            default: 0
        - name: count
          in: query
          description: Data count
          required: true
          explode: true
          schema:
            type: number
            default: 5
        - name: patientsId
          in: query
          description: Patients Id
          required: false
          explode: true
          schema:
            type: number
            default: 4598
        - name: phoneNumber
          in: query
          description: Patients Phone Number
          required: false
          explode: true
          schema:
            type: string
            default: 7010182967
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Patients'
        '400':
          description: Invalid user id
  /v1/api/patients/{id}:
    get:
      tags:
        - Patients
      summary: Creation of new Patients
      description: Return the list values
      operationId: getPatientsById
      parameters:
        - name: id
          in: path
          description: User Id
          required: true
          explode: true
          schema:
            type: string
            default: available
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Patients'
        '400':
          description: Invalid user id
    put:
      tags:
        - Patients
      summary: Creation of new patients
      description: Return the list values
      operationId: updatePatients
      requestBody:
        description: Update an existent Patients
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Patients'
        required: true
      parameters:
        - name: id
          in: path
          description: Patients Id
          required: true
          explode: true
          schema:
            type: string
            default: available
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Patients'
        '400':
          description: Invalid Patients id
    delete:
      tags:
        - Patients
      summary: Delete Operation
      description: Delete patients account
      operationId: deletePatientsById
      parameters:
        - name: id
          in: path
          description: User Id
          required: true
          explode: true
          schema:
            type: string
            default: available
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: string
                example: "Doctor removed successfully"
        '400':
          description: Invalid user id
...
components:
  schemas:
...
    UserBasiInfo:
      type: object
      properties:
        user_id:
          type: string
          example: fb17eecf-5ec6-4c13-a5bd-5ce1451472d0
        first_name:
          type: string
          example: Sastika
        last_name:
          type: string
          example: Aravind
        role:
          $ref: '#/components/schemas/Role'
        gender:
          type: string
          example: Male
          enum:
            - Male
            - FeMale
            - Others
        employment_details:
          $ref: '#/components/schemas/EmploymentDetails'
    Address:
      type: object
      properties:
        id:
          type: string
          example: 4acd0b30-0ac1-4d0f-938b-c7b86d804375
        address_category:
          type: string
          example: Personal
          enum:
            - CURRENT
            - OFFICE
            - PERMANENT
        country:
          type: string
          example: India
        state:
          type: string
          example: Tamilnadu
        city:
          type: string
          example: Coimabtore
        street:
          type: string
          example: 121,Balaji Nagar
        pin:
          type: string
          example: 624601
        geo_location:
          type: string
          example: India
    Contact:
      type: object
      properties:
        id:
          type: string
          example: 4acd0b30-0ac1-4d0f-938b-c7b86d804375
        address:
          $ref: '#/components/schemas/Address'
        phone:
          type: string
          example: 7010192967
        email:
          type: string
          example: aravind@gmail.com
...
    Patients:
      type: object
      properties:
        id:
          type: string
          example: 4acd0b30-0ac1-4d0f-938b-c7b86d804375
        basic_info:
          $ref: '#/components/schemas/UserBasiInfo'
        contact:
          $ref: '#/components/schemas/Contact'
...
    Role:
      type: object
      properties:
        id:
          type: string
          example: 4acd0b30-0ac1-4d0f-938b-c7b86d804375
        name:
          type: string
          example: ROLE_USER
        description:
          type: string
          example: "User role"
    EmploymentDetails:
      type: object
      properties:
        id:
          type: string
          example: 4acd0b30-0ac1-4d0f-938b-c7b86d804375
        joining_date:
          type: string
          format: date-time
          pattern: "^(\\d{4})-(\\d{2})-(\\d{2})T((\\d{2}):(\\d{2}):(\\d{2}))\\.(\\d{3})Z$"
          description: "User joining date."
          example: "2024-11-05T00:00:00Z"
        relieving_date:
          type: string
          format: date-time
          pattern: "^(\\d{4})-(\\d{2})-(\\d{2})T((\\d{2}):(\\d{2}):(\\d{2}))\\.(\\d{3})Z$"
          description: "User joining date."
          example: "2024-11-05T00:00:00Z"
        department:
          type: string
          example: "CLINIC_STAFF"
        salary:
          type: number
          example: 18500
        payroll:
          $ref: '#/components/schemas/PayRoll'
    PayRoll:
      type: object
      properties:
        id:
          type: string
          example: 4acd0b30-0ac1-4d0f-938b-c7b86d804375
        user:
          $ref: '#/components/schemas/UserBasiInfo'
        payment_date:
          type: string
          format: date-time
          pattern: "^(\\d{4})-(\\d{2})-(\\d{2})T((\\d{2}):(\\d{2}):(\\d{2}))\\.(\\d{3})Z$"
          description: "Appointment Booking Date"
          example: "2024-11-05T00:00:00Z"
        gross_amount:
          type: number
          format: float
        deductions:
          type: number
          format: float
        net_amount:
          type: number
          format: float
...